+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 0.1.1dev
-
+* Adds `pkgmt check-project`
 ## 0.1.0 (2023-01-06)
 * Completing version string when missing components. (e.g., `1.0` -> `1.0.0`) ([#18](https://github.com/ploomber/pkgmt/issues/18))